pyproject.toml
[project]
name = "chapterly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg2-binary>=2.9.10",
    "django-environ>=0.11.2",
    "pillow>=11.1.0",
    "django-filter>=24.3",
    "drf-spectacular>=0.27.2",
]
